Pinta mas de una esfera

+++ src/get_data/create_objects.c
@@ -71,15 +71,15 @@ t_light	create_light(char *s, t_world *all)
 		s++;
 	s = ft_strchr(s, ' ');
 
-	new.rgb.r = ft_atoi(++s) / 255;
+	new.rgb.r = (double)ft_atoi(++s) / 255;
 	new.intensity.r = new.rgb.r * new.brightness ;
 	while (s && *s != ',')
 		s++;
-	new.rgb.g = ft_atoi(++s) / 255;
+	new.rgb.g = (double)ft_atoi(++s) / 255;
 	new.intensity.g = new.rgb.g * new.brightness;
 	while (s && *s != ',')
 		s++;
-	new.rgb.b = ft_atoi(++s) / 255;
+	new.rgb.b = (double)ft_atoi(++s) / 255;
 	new.intensity.b = new.rgb.b * new.brightness;
 	return new;
 }
